+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New features
 
 * Service container for internal services
 * Set /Lang entry for better accessibility when document language is available (@cuongmits, #1418)
+* More verbose helper methods for `Output`: `OutputBinaryData`, `OutputHttpInline`, `OutputHttpDownload`, `OutputFile`
 
 Bugfixes
 --------